+++ docs/fr_FR/Changelog.md
@@ -4,6 +4,8 @@
 -   Améliorations: Modéles: Suppression cmd obsolete 'setReportLightColorSpectre' => 'act_zbConfigureReporting2'
 -   Améliorations: Modéles: Suppression cmd obsolete 'setReportPuissanceLegrand' => 'act_zbConfigureReporting2'
 -   Correction: Erreur interne cmd 'setReport'.
+-   Amélioration: Interne cmd.
+-   Amélioration interne cmd: checkZigatesStatus() => intégré dans processCmdQueues()
 
 ## 240920-BETA-3